fix: address anonymize runtime review comments
