+++ apps/web/src/i18n/langs/cs.json
@@ -541,7 +541,7 @@
     "anonymizationLabels": {
       "address": "Adresa",
       "bankAccountNumber": "Číslo bankovního účtu",
-      "birthNumber": "Birth number",
+      "birthNumber": "Rodné číslo",
       "creditCardNumber": "Číslo platební karty",
       "date": "Datum",
       "dateOfBirth": "Datum narození",
@@ -551,13 +551,13 @@
       "landParcel": "Parcela",
       "miscellaneous": "Ostatní",
       "monetaryAmount": "Peněžní částka",
-      "nationalIdentificationNumber": "National identification number",
+      "nationalIdentificationNumber": "Národní identifikační číslo",
       "organization": "Organizace",
       "passportNumber": "Číslo pasu",
       "person": "Osoba",
       "phoneNumber": "Telefonní číslo",
       "registrationNumber": "IČO",
-      "socialSecurityNumber": "Social security number",
+      "socialSecurityNumber": "Číslo sociálního pojištění",
       "taxIdentificationNumber": "DIČ"
     },
     "cancel": "Zrušit",
