+++ apps/web/src/i18n/langs/de.json
@@ -541,7 +541,7 @@
     "anonymizationLabels": {
       "address": "Adresse",
       "bankAccountNumber": "Bankkontonummer",
-      "birthNumber": "Birth number",
+      "birthNumber": "Geburtsnummer",
       "creditCardNumber": "Kreditkartennummer",
       "date": "Datum",
       "dateOfBirth": "Geburtsdatum",
@@ -551,13 +551,13 @@
       "landParcel": "Flurstück",
       "miscellaneous": "Sonstiges",
       "monetaryAmount": "Geldbetrag",
-      "nationalIdentificationNumber": "National identification number",
+      "nationalIdentificationNumber": "Nationale Identifikationsnummer",
       "organization": "Organisation",
       "passportNumber": "Reisepassnummer",
       "person": "Person",
       "phoneNumber": "Telefonnummer",
       "registrationNumber": "Registernummer",
-      "socialSecurityNumber": "Social security number",
+      "socialSecurityNumber": "Sozialversicherungsnummer",
       "taxIdentificationNumber": "Steuer-Identifikationsnummer"
     },
     "cancel": "Abbrechen",
