+++ apps/web/src/i18n/langs/es.json
@@ -541,7 +541,7 @@
     "anonymizationLabels": {
       "address": "Dirección",
       "bankAccountNumber": "Número de cuenta bancaria",
-      "birthNumber": "Birth number",
+      "birthNumber": "Número de nacimiento",
       "creditCardNumber": "Número de tarjeta de crédito",
       "date": "Fecha",
       "dateOfBirth": "Fecha de nacimiento",
@@ -551,13 +551,13 @@
       "landParcel": "Parcela",
       "miscellaneous": "Otros",
       "monetaryAmount": "Importe",
-      "nationalIdentificationNumber": "National identification number",
+      "nationalIdentificationNumber": "Número nacional de identificación",
       "organization": "Organización",
       "passportNumber": "Número de pasaporte",
       "person": "Persona",
       "phoneNumber": "Número de teléfono",
       "registrationNumber": "Número de registro",
-      "socialSecurityNumber": "Social security number",
+      "socialSecurityNumber": "Número de la seguridad social",
       "taxIdentificationNumber": "NIF / CIF"
     },
     "cancel": "Cancelar",
