+++ apps/web/src/i18n/langs/et.json
@@ -541,7 +541,7 @@
     "anonymizationLabels": {
       "address": "Aadress",
       "bankAccountNumber": "Pangakonto number",
-      "birthNumber": "Birth number",
+      "birthNumber": "Sünninumber",
       "creditCardNumber": "Krediitkaardi number",
       "date": "Kuupäev",
       "dateOfBirth": "Sünnikuupäev",
@@ -551,13 +551,13 @@
       "landParcel": "Maatükk",
       "miscellaneous": "Muu",
       "monetaryAmount": "Rahasumma",
-      "nationalIdentificationNumber": "National identification number",
+      "nationalIdentificationNumber": "Riiklik isikukood",
       "organization": "Organisatsioon",
       "passportNumber": "Passi number",
       "person": "Isik",
       "phoneNumber": "Telefoninumber",
       "registrationNumber": "Registrikood",
-      "socialSecurityNumber": "Social security number",
+      "socialSecurityNumber": "Sotsiaalkindlustuse number",
       "taxIdentificationNumber": "Maksukohustuslase number"
     },
     "cancel": "Tühista",
