+++ apps/web/src/i18n/langs/fr.json
@@ -541,7 +541,7 @@
     "anonymizationLabels": {
       "address": "Adresse",
       "bankAccountNumber": "Numéro de compte bancaire",
-      "birthNumber": "Birth number",
+      "birthNumber": "Numéro de naissance",
       "creditCardNumber": "Numéro de carte bancaire",
       "date": "Date",
       "dateOfBirth": "Date de naissance",
@@ -551,13 +551,13 @@
       "landParcel": "Parcelle",
       "miscellaneous": "Divers",
       "monetaryAmount": "Montant",
-      "nationalIdentificationNumber": "National identification number",
+      "nationalIdentificationNumber": "Numéro national d’identification",
       "organization": "Organisation",
       "passportNumber": "Numéro de passeport",
       "person": "Personne",
       "phoneNumber": "Numéro de téléphone",
       "registrationNumber": "Numéro d’enregistrement",
-      "socialSecurityNumber": "Social security number",
+      "socialSecurityNumber": "Numéro de sécurité sociale",
       "taxIdentificationNumber": "Numéro fiscal"
     },
     "cancel": "Annuler",
