+++ apps/web/src/i18n/langs/hu.json
@@ -541,7 +541,7 @@
     "anonymizationLabels": {
       "address": "Cím",
       "bankAccountNumber": "Bankszámlaszám",
-      "birthNumber": "Birth number",
+      "birthNumber": "Születési szám",
       "creditCardNumber": "Bankkártyaszám",
       "date": "Dátum",
       "dateOfBirth": "Születési dátum",
@@ -551,13 +551,13 @@
       "landParcel": "Földrészlet",
       "miscellaneous": "Egyéb",
       "monetaryAmount": "Pénzösszeg",
-      "nationalIdentificationNumber": "National identification number",
+      "nationalIdentificationNumber": "Nemzeti azonosító szám",
       "organization": "Szervezet",
       "passportNumber": "Útlevélszám",
       "person": "Személy",
       "phoneNumber": "Telefonszám",
       "registrationNumber": "Cégjegyzékszám",
-      "socialSecurityNumber": "Social security number",
+      "socialSecurityNumber": "Társadalombiztosítási szám",
       "taxIdentificationNumber": "Adóazonosító"
     },
     "cancel": "Mégse",
