+++ apps/web/src/i18n/langs/lt.json
@@ -541,7 +541,7 @@
     "anonymizationLabels": {
       "address": "Adresas",
       "bankAccountNumber": "Banko sąskaitos numeris",
-      "birthNumber": "Birth number",
+      "birthNumber": "Gimimo numeris",
       "creditCardNumber": "Kredito kortelės numeris",
       "date": "Data",
       "dateOfBirth": "Gimimo data",
@@ -551,13 +551,13 @@
       "landParcel": "Sklypas",
       "miscellaneous": "Kita",
       "monetaryAmount": "Pinigų suma",
-      "nationalIdentificationNumber": "National identification number",
+      "nationalIdentificationNumber": "Nacionalinis identifikavimo numeris",
       "organization": "Organizacija",
       "passportNumber": "Paso numeris",
       "person": "Asmuo",
       "phoneNumber": "Telefono numeris",
       "registrationNumber": "Įmonės kodas",
-      "socialSecurityNumber": "Social security number",
+      "socialSecurityNumber": "Socialinio draudimo numeris",
       "taxIdentificationNumber": "Mokesčių mokėtojo kodas"
     },
     "cancel": "Atšaukti",
