+++ apps/web/src/i18n/langs/lv.json
@@ -541,7 +541,7 @@
     "anonymizationLabels": {
       "address": "Adrese",
       "bankAccountNumber": "Bankas konta numurs",
-      "birthNumber": "Birth number",
+      "birthNumber": "Dzimšanas numurs",
       "creditCardNumber": "Kredītkartes numurs",
       "date": "Datums",
       "dateOfBirth": "Dzimšanas datums",
@@ -551,13 +551,13 @@
       "landParcel": "Zemes vienība",
       "miscellaneous": "Cits",
       "monetaryAmount": "Naudas summa",
-      "nationalIdentificationNumber": "National identification number",
+      "nationalIdentificationNumber": "Nacionālais identifikācijas numurs",
       "organization": "Organizācija",
       "passportNumber": "Pases numurs",
       "person": "Persona",
       "phoneNumber": "Tālruņa numurs",
       "registrationNumber": "Reģistrācijas numurs",
-      "socialSecurityNumber": "Social security number",
+      "socialSecurityNumber": "Sociālās apdrošināšanas numurs",
       "taxIdentificationNumber": "Nodokļu maksātāja numurs"
     },
     "cancel": "Atcelt",
