+++ apps/web/src/i18n/langs/pl.json
@@ -541,7 +541,7 @@
     "anonymizationLabels": {
       "address": "Adres",
       "bankAccountNumber": "Numer rachunku bankowego",
-      "birthNumber": "Birth number",
+      "birthNumber": "Numer urodzenia",
       "creditCardNumber": "Numer karty kredytowej",
       "date": "Data",
       "dateOfBirth": "Data urodzenia",
@@ -551,13 +551,13 @@
       "landParcel": "Działka",
       "miscellaneous": "Inne",
       "monetaryAmount": "Kwota",
-      "nationalIdentificationNumber": "National identification number",
+      "nationalIdentificationNumber": "Krajowy numer identyfikacyjny",
       "organization": "Organizacja",
       "passportNumber": "Numer paszportu",
       "person": "Osoba",
       "phoneNumber": "Numer telefonu",
       "registrationNumber": "Numer rejestrowy",
-      "socialSecurityNumber": "Social security number",
+      "socialSecurityNumber": "Numer ubezpieczenia społecznego",
       "taxIdentificationNumber": "NIP"
     },
     "cancel": "Anuluj",
