+++ apps/web/src/i18n/langs/pt-BR.json
@@ -541,7 +541,7 @@
     "anonymizationLabels": {
       "address": "Endereço",
       "bankAccountNumber": "Número da conta bancária",
-      "birthNumber": "Birth number",
+      "birthNumber": "Número de nascimento",
       "creditCardNumber": "Número do cartão de crédito",
       "date": "Data",
       "dateOfBirth": "Data de nascimento",
@@ -551,13 +551,13 @@
       "landParcel": "Lote",
       "miscellaneous": "Outros",
       "monetaryAmount": "Valor monetário",
-      "nationalIdentificationNumber": "National identification number",
+      "nationalIdentificationNumber": "Número nacional de identificação",
       "organization": "Organização",
       "passportNumber": "Número do passaporte",
       "person": "Pessoa",
       "phoneNumber": "Número de telefone",
       "registrationNumber": "CNPJ",
-      "socialSecurityNumber": "Social security number",
+      "socialSecurityNumber": "Número de seguro social",
       "taxIdentificationNumber": "CPF / CNPJ"
     },
     "cancel": "Cancelar",
