+++ apps/web/src/i18n/langs/sk.json
@@ -541,7 +541,7 @@
     "anonymizationLabels": {
       "address": "Adresa",
       "bankAccountNumber": "Číslo bankového účtu",
-      "birthNumber": "Birth number",
+      "birthNumber": "Rodné číslo",
       "creditCardNumber": "Číslo platobnej karty",
       "date": "Dátum",
       "dateOfBirth": "Dátum narodenia",
@@ -551,13 +551,13 @@
       "landParcel": "Parcela",
       "miscellaneous": "Ostatné",
       "monetaryAmount": "Peňažná suma",
-      "nationalIdentificationNumber": "National identification number",
+      "nationalIdentificationNumber": "Národné identifikačné číslo",
       "organization": "Organizácia",
       "passportNumber": "Číslo pasu",
       "person": "Osoba",
       "phoneNumber": "Telefónne číslo",
       "registrationNumber": "IČO",
-      "socialSecurityNumber": "Social security number",
+      "socialSecurityNumber": "Číslo sociálneho poistenia",
       "taxIdentificationNumber": "DIČ"
     },
     "cancel": "Zrušiť",
